Fixed bug on getOne Added 404 on delete not existing entity Added CORS Added test index.html with FETCH requests

+++ src/modules/todo/controllers/todo.controller.ts
@@ -23,7 +23,7 @@ export class TodoController {
 
   @Get(':id')
   async getOneAction(@Param('id') id: string): Promise<Todo> {
-    const todo = this.todoService.findOne(id);
+    const todo = await this.todoService.findOne(id);
     if (todo === undefined) {
       throw new HttpException(
         'Todo with id=' + id + ' not exists',
@@ -58,7 +58,14 @@ export class TodoController {
   }
 
   @Delete(':id')
-  deleteAction(@Param('id') id: string): Promise<void> {
-    return this.todoService.remove(id);
+  async deleteAction(@Param('id') id: string): Promise<{success: boolean}> {
+    const todo = await this.todoService.findOne(id);
+    if (todo === undefined) {
+      throw new NotFoundException('Todo with id=' + id + ' not exists');
+    }
+    await this.todoService.remove(id);
+    return {
+      success: true
+    };
   }
 }
